AddedLoginToDeleteEndpoint

+++ src/main/java/com/example/data22aexamprojectlnl2/controllers/ImageController.java
@@ -1,7 +1,10 @@
 package com.example.data22aexamprojectlnl2.controllers;
 
+import com.example.data22aexamprojectlnl2.models.Company;
 import com.example.data22aexamprojectlnl2.models.Image;
+import com.example.data22aexamprojectlnl2.services.CompanyService;
 import com.example.data22aexamprojectlnl2.services.ImageService;
+import com.example.data22aexamprojectlnl2.services.PasswordHashingService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,17 +17,30 @@ import java.util.Optional;
 @Controller
 public class ImageController {
 
+
+    final PasswordHashingService passwordHashing = new PasswordHashingService();
     @Autowired
     ImageService imageService;
+    @Autowired
+    CompanyService companyService;
 
     @DeleteMapping("/deleteImage")
-    public ResponseEntity<String> deleteImageById(@RequestParam("image_id") int image_id){
-        Optional<Image> foundImage = imageService.getImageById(image_id);
-        if(foundImage.isPresent()) {
-            imageService.deleteImage(image_id);
-            return ResponseEntity.ok("Image was deleted");
-        } else {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Image was not found");
+    public ResponseEntity<String> deleteImageById(@RequestParam("image_id") int image_id,
+                                                  @RequestParam("username") String username,
+                                                  @RequestParam("password") String password){
+        String hashedUsername = passwordHashing.doHashing(username);
+        String hashedPassword = passwordHashing.doHashing(password);
+        Optional<Company> checkCompany = companyService.getCompanyByUsernameAndPassword(hashedUsername, hashedPassword);
+        if(checkCompany.isPresent()) {
+            Optional<Image> foundImage = imageService.getImageById(image_id);
+            if (foundImage.isPresent()) {
+                imageService.deleteImage(image_id);
+                return ResponseEntity.ok("Image was deleted");
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Image was not found");
+            }
+        }else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Wrong username or password");
         }
     }
 }
