+++ src/main/java/com/example/data22aexamprojectlnl2/controllers/PosterController.java
@@ -1,9 +1,12 @@
 package com.example.data22aexamprojectlnl2.controllers;
 
+import com.example.data22aexamprojectlnl2.models.Company;
 import com.example.data22aexamprojectlnl2.models.Image;
 import com.example.data22aexamprojectlnl2.models.Poster;
 import com.example.data22aexamprojectlnl2.repositories.ImageRepository;
+import com.example.data22aexamprojectlnl2.services.CompanyService;
 import com.example.data22aexamprojectlnl2.services.ImageService;
+import com.example.data22aexamprojectlnl2.services.PasswordHashingService;
 import com.example.data22aexamprojectlnl2.services.PosterService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -22,12 +25,16 @@ import java.util.Optional;
 @Controller
 public class PosterController
 {
+
+    final PasswordHashingService passwordHashing = new PasswordHashingService();
     @Autowired
     ImageService imageService;
     @Autowired
     ImageRepository imageRepository;
     @Autowired
     PosterService posterService;
+    @Autowired
+    CompanyService companyService;
 
     @GetMapping("/getImages")
     public ResponseEntity<List<Image>> getAllImagesByPosterId(@RequestParam("poster_id") int poster_id)
@@ -60,13 +67,22 @@ public class PosterController
     } //
 
     @DeleteMapping("/deletePoster")
-    public ResponseEntity<String> deleterPosterByPosterId(@RequestParam("poster_id") int poster_id) {
-        Optional<Poster> checkPoster = posterService.getPosterById(poster_id);
-        if(checkPoster.isPresent()) {
-            posterService.deletePoster(poster_id);
-            return ResponseEntity.ok("Poster deleted");
+    public ResponseEntity<String> deleterPosterByPosterId(@RequestParam("poster_id") int poster_id,
+                                                          @RequestParam("username") String username,
+                                                          @RequestParam("password") String password) {
+        String hashedUsername = passwordHashing.doHashing(username);
+        String hashedPassword = passwordHashing.doHashing(password);
+        Optional<Company> checkCompany = companyService.getCompanyByUsernameAndPassword(hashedUsername, hashedPassword);
+        if(checkCompany.isPresent()) {
+            Optional<Poster> checkPoster = posterService.getPosterById(poster_id);
+            if (checkPoster.isPresent()) {
+                posterService.deletePoster(poster_id);
+                return ResponseEntity.ok("Poster deleted");
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Poster not found");
+            }
         } else {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Poster not found");
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Wrong username or password");
         }
     }
